Notify the user that an eject has occurred
This change will blink the LED and buzz the optional piezo speaker
a visible and audible signal for about a second when the device
registers an eject. It is configurable via 'zuluscsi.ini' using
`EjectBlinkTimes`
`EjectBlinkPeriod`
under `[SCSI]` or per device under `[SCSIx]` where x is the SCSI
device ID.

```
EjectBlinkTimes = 20
EjectBlinkPeriod = 50
```
are the default values and a good starting point for customization.
