+++ zuluscsi.ini
@@ -74,6 +74,11 @@
 #ReinsertCDOnInquiry = 1 # Reinsert any ejected CD-ROM image on Inquiry command
 #ReinsertAfterEject = 1 # Reinsert next CD image after eject, if multiple images configured.
 #EjectButton = 0 # Enable eject by button 1 or 2, or set 0 to disable
+
+# The default values of 20 and 50 give a visible and audible (with optional buzzer) eject notification
+#EjectBlinkTimes = 20 # the number of times to blink the LED
+#EjectBlinkPeriod = 50 # the period in milliseconds for each blink
+
 #CDAVolume = 63 # Change CD Audio default volume. Maximum 255.
 #DisableMacSanityCheck = 0 # Disable sanity warnings for Mac disk drives. Default is 0 - enable checks
 #BlockSize = 0 # Set the drive's blocksize, defaults to 2048 for CDs and 512 for all other drives
